requirements.txt
attr==0.3.2
attrs==22.2.0
fastapi==0.88.0
fastapi_login==1.8.3
joblib==1.2.0
motor==3.1.1
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.0
requests==2.28.1
twilio==7.16.0
uvicorn==0.20.0
